requirements.txt
torch
adapter-transformers
ftfy
tqdm
sklearn